Consistent spelling of "Signalling"
