+++ Samples/DataChannelOrtc/DataChannelOrtc/MainPage.xaml.cs
@@ -19,7 +19,7 @@ using Windows.UI.Xaml.Data;
 using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
-using DataChannelOrtc.Signalling;
+using DataChannelOrtc.Signaling;
 
 // The Blank Page item template is documented at http://go.microsoft.com/fwlink/?LinkId=402352&clcid=0x409
