+++ Samples/PeerCC-Sample/Signaling/Signaling.cs
@@ -19,36 +19,68 @@ using Windows.Networking;
 using Windows.Networking.Sockets;
 using Windows.Storage.Streams;
 
-namespace DataChannelOrtc.Signalling
+namespace PeerConnectionClient.Signaling
 {
+    public delegate void SignedInDelegate();
+    public delegate void DisconnectedDelegate();
+    public delegate void PeerConnectedDelegate(int id, string name);
+    public delegate void PeerDisonnectedDelegate(int peer_id);
+    public delegate void PeerHangupDelegate(int peer_id);
+    public delegate void MessageFromPeerDelegate(int peer_id, string message);
+    public delegate void MessageSentDelegate(int err);
+    public delegate void ServerConnectionFailureDelegate();
+
     /// <summary>
-    /// Signaller instance is used to fire connection events.
+    /// Signaler instance is used to fire connection events.
     /// </summary>
-    class TcpSignaler : Signaler
+    class Signaler
     {
+        // Connection events
+        public event SignedInDelegate OnSignedIn;
+        public event DisconnectedDelegate OnDisconnected;
+        public event PeerConnectedDelegate OnPeerConnected;
+        public event PeerDisonnectedDelegate OnPeerDisconnected;
+        public event PeerHangupDelegate OnPeerHangup;
+        public event MessageFromPeerDelegate OnMessageFromPeer;
+        public event ServerConnectionFailureDelegate OnServerConnectionFailure;
+
+        /// <summary>
+        /// Creates an instance of a Signaler.
+        /// </summary>
+        public Signaler()
+        {
+            _state = State.NOT_CONNECTED;
+            _myId = -1;
+
+            // Annoying but register empty handlers
+            // so we don't have to check for null everywhere
+            OnSignedIn += () => { };
+            OnDisconnected += () => { };
+            OnPeerConnected += (a, b) => { };
+            OnPeerDisconnected += (a) => { };
+            OnMessageFromPeer += (a, b) => { };
+            OnServerConnectionFailure += () => { };
+        }
+
+        /// <summary>
+        /// The connection state.
+        /// </summary>
+        public enum State
+        {
+            NOT_CONNECTED,
+            RESOLVING, // Note: State not used
+            SIGNING_IN,
+            CONNECTED,
+            SIGNING_OUT_WAITING, // Note: State not used
+            SIGNING_OUT,
+        };
         private State _state;
 
         private HostName _server;
         private string _port;
         private string _clientName;
         private int _myId;
-
         private Dictionary<int, string> _peers = new Dictionary<int, string>();
-        private StreamSocket _hangingGetSocket;
-
-        /// <summary>
-        /// Creates an instance of a Signaller.
-        /// </summary>
-        public TcpSignaler(string server, string port, string clientName)
-        {
-            _state = State.NOT_CONNECTED;
-            _myId = -1;
-
-            _server = new HostName(server);
-            _port = port;
-            _clientName = clientName;
-        }
-
 
         /// <summary>
         /// Checks if connected to the server.
@@ -65,18 +97,22 @@ namespace DataChannelOrtc.Signalling
         /// <param name="server">Host name/IP.</param>
         /// <param name="port">Port to connect.</param>
         /// <param name="client_name">Client name.</param>
-        public override async Task Connect()
+        public async void Connect(string server, string port, string client_name)
         {
             try
             {
                 if (_state != State.NOT_CONNECTED)
                 {
-                    OnConnectionFailed();
+                    OnServerConnectionFailure();
                     return;
                 }
 
+                _server = new HostName(server);
+                _port = port;
+                _clientName = client_name;
+
                 _state = State.SIGNING_IN;
-                await ControlSocketRequestAsync(string.Format("GET /sign_in?{0} HTTP/1.0\r\n\r\n", _clientName));
+                await ControlSocketRequestAsync(string.Format("GET /sign_in?{0} HTTP/1.0\r\n\r\n", client_name));
                 if (_state == State.CONNECTED)
                 {
                     // Start the long polling loop without await
@@ -85,7 +121,7 @@ namespace DataChannelOrtc.Signalling
                 else
                 {
                     _state = State.NOT_CONNECTED;
-                    OnConnectionFailed();
+                    OnServerConnectionFailure();
                 }
             }
             catch (Exception ex)
@@ -94,94 +130,7 @@ namespace DataChannelOrtc.Signalling
             }
         }
 
-        /// <summary>
-        /// Disconnects the user from the server.
-        /// </summary>
-        /// <returns>True if the user is disconnected from the server.</returns>
-        public async Task<bool> SignOut()
-        {
-            if (_state == State.NOT_CONNECTED || _state == State.SIGNING_OUT)
-                return true;
-
-            if (_hangingGetSocket != null)
-            {
-                _hangingGetSocket.Dispose();
-                _hangingGetSocket = null;
-            }
-
-            _state = State.SIGNING_OUT;
-
-            if (_myId != -1)
-            {
-                await ControlSocketRequestAsync(String.Format("GET /sign_out?peer_id={0} HTTP/1.0\r\n\r\n", _myId));
-            }
-            else
-            {
-                // Can occur if the app is closed before we finish connecting
-                return true;
-            }
-
-            _myId = -1;
-            _state = State.NOT_CONNECTED;
-            return true;
-        }
-
-        /// <summary>
-        /// Resets the states after connection is closed.
-        /// </summary>
-        private void Close()
-        {
-            if (_hangingGetSocket != null)
-            {
-                _hangingGetSocket.Dispose();
-                _hangingGetSocket = null;
-            }
-
-            _peers.Clear();
-            _state = State.NOT_CONNECTED;
-        }
-
-        /// <summary>
-        /// Sends a message to a peer.
-        /// </summary>
-        /// <param name="peerId">ID of the peer to send a message to.</param>
-        /// <param name="message">Message to send.</param>
-        /// <returns>True if the message was sent.</returns>
-        public override async Task<bool> SendToPeer(int peerId, string message)
-        {
-            if (_state != State.CONNECTED)
-            {
-                return false;
-            }
-
-            Debug.Assert(IsConnected());
-
-            if (!IsConnected() || peerId == -1)
-            {
-                return false;
-            }
-
-            string buffer = String.Format(
-                "POST /message?peer_id={0}&to={1} HTTP/1.0\r\n" +
-                "Content-Length: {2}\r\n" +
-                "Content-Type: text/plain\r\n" +
-                "\r\n" +
-                "{3}",
-                _myId, peerId, message.Length, message);
-            return await ControlSocketRequestAsync(buffer);
-        }
-
-        /// <summary>
-        /// Sends a message to a peer.
-        /// </summary>
-        /// <param name="peerId">ID of the peer to send a message to.</param>
-        /// <param name="json">The json message.</param>
-        /// <returns>True if the message is sent.</returns>
-        public async Task<bool> SendToPeer(int peerId, IJsonValue json)
-        {
-            string message = json.Stringify();
-            return await SendToPeer(peerId, message);
-        }
+        private StreamSocket _hangingGetSocket;
 
         #region Parsing
         /// <summary>
@@ -275,7 +224,7 @@ namespace DataChannelOrtc.Signalling
                     {
                         Debug.WriteLine("Peer most likely gone. Closing peer connection.");
                         //As Peer Id doesn't exist in buffer using 0
-                        OnPeerDisconnected(new Signalling.Peer(0, string.Empty));
+                        OnPeerDisconnected(0);
                         return false;
                     }
                     Close();
@@ -469,11 +418,11 @@ namespace DataChannelOrtc.Signalling
                             if (ParseEntry(buffer.Substring(pos, eol - pos), ref name, ref id, ref connected) && id != _myId)
                             {
                                 _peers[id] = name;
-                                OnPeerConnected(new Peer(id, name));
+                                OnPeerConnected(id, name);
                             }
                             pos = eol + 1;
                         }
-                        OnConnected();
+                        OnSignedIn();
                     }
                 }
                 else if (_state == State.SIGNING_OUT)
@@ -547,12 +496,12 @@ namespace DataChannelOrtc.Signalling
                                 if (connected)
                                 {
                                     _peers[id] = name;
-                                    OnPeerConnected(new Peer(id, name));
+                                    OnPeerConnected(id, name);
                                 }
                                 else
                                 {
                                     _peers.Remove(id);
-                                    OnPeerDisconnected(new Peer(id, string.Empty));
+                                    OnPeerDisconnected(id);
                                 }
                             }
                         }
@@ -561,21 +510,134 @@ namespace DataChannelOrtc.Signalling
                             string message = buffer.Substring(pos);
                             if (message == "BYE")
                             {
-                                OnPeerHangup(new Peer(peer_id, string.Empty));
+                                OnPeerHangup(peer_id);
                             }
                             else
                             {
-                                OnPeerMessage(new Peer(peer_id, string.Empty, message));
+                                OnMessageFromPeer(peer_id, message);
                             }
                         }
                     }
                     catch (Exception e)
                     {
-                        // Below is noisy during debugging.
-                        //Debug.WriteLine("[Error] Signaling: Long-polling exception: {0}", e.Message);
+                        Debug.WriteLine("[Error] Signaling: Long-polling exception: {0}", e.Message);
                     }
                 }
             }
+        }
+
+        /// <summary>
+        /// Disconnects the user from the server.
+        /// </summary>
+        /// <returns>True if the user is disconnected from the server.</returns>
+        public async Task<bool> SignOut()
+        {
+            if (_state == State.NOT_CONNECTED || _state == State.SIGNING_OUT)
+                return true;
+
+            if (_hangingGetSocket != null)
+            {
+                _hangingGetSocket.Dispose();
+                _hangingGetSocket = null;
+            }
+
+            _state = State.SIGNING_OUT;
+
+            if (_myId != -1)
+            {
+                await ControlSocketRequestAsync(String.Format("GET /sign_out?peer_id={0} HTTP/1.0\r\n\r\n", _myId));
+            }
+            else
+            {
+                // Can occur if the app is closed before we finish connecting
+                return true;
+            }
+
+            _myId = -1;
+            _state = State.NOT_CONNECTED;
+            return true;
+        }
+
+        /// <summary>
+        /// Resets the states after connection is closed.
+        /// </summary>
+        private void Close()
+        {
+            if (_hangingGetSocket != null)
+            {
+                _hangingGetSocket.Dispose();
+                _hangingGetSocket = null;
+            }
+
+            _peers.Clear();
+            _state = State.NOT_CONNECTED;
+        }
+
+        /// <summary>
+        /// Sends a message to a peer.
+        /// </summary>
+        /// <param name="peerId">ID of the peer to send a message to.</param>
+        /// <param name="message">Message to send.</param>
+        /// <returns>True if the message was sent.</returns>
+        public async Task<bool> SendToPeer(int peerId, string message)
+        {
+            if (_state != State.CONNECTED)
+            {
+                return false;
+            }
+
+            Debug.Assert(IsConnected());
+
+            if (!IsConnected() || peerId == -1)
+            {
+                return false;
+            }
+
+            string buffer = String.Format(
+                "POST /message?peer_id={0}&to={1} HTTP/1.0\r\n" +
+                "Content-Length: {2}\r\n" +
+                "Content-Type: text/plain\r\n" +
+                "\r\n" +
+                "{3}",
+                _myId, peerId, message.Length, message);
+            return await ControlSocketRequestAsync(buffer);
+        }
+
+        /// <summary>
+        /// Sends a message to a peer.
+        /// </summary>
+        /// <param name="peerId">ID of the peer to send a message to.</param>
+        /// <param name="json">The json message.</param>
+        /// <returns>True if the message is sent.</returns>
+        public async Task<bool> SendToPeer(int peerId, IJsonValue json)
+        {
+            string message = json.Stringify();
+            return await SendToPeer(peerId, message);
+        }
+    }
+
+    /// <summary>
+    /// Class providing helper functions for parsing responses and messages.
+    /// </summary>
+    public static class Extensions
+    {
+        public static async void WriteStringAsync(this StreamSocket socket, string str)
+        {
+            try
+            {
+                var writer = new DataWriter(socket.OutputStream);
+                writer.WriteString(str);
+                await writer.StoreAsync();
+            }
+            catch (Exception ex)
+            {
+                Debug.WriteLine("[Error] Singnaling: Couldn't write to socket : " + ex.Message);
+            }
+        }
+
+        public static int ParseLeadingInt(this string str)
+        {
+            return int.Parse(Regex.Match(str, "\\d+").Value);
         }
     }
 }
